test(ProdutosService): sergregando as responsabilidades
